feat: separa CORS de dev e prod (Ref. #277)
